backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
alembic
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pydantic[email]
email-validator